reorganize constants folder
