+++ src/constants/config.js
@@ -1,4 +1,9 @@
-export const API_URL = 'http://localhost:3000/api'
 export const CONFIG_REQUEST = 'CONFIG_REQUEST'
 export const CONFIG_SUCCESS = 'CONFIG_SUCCESS'
 export const CONFIG_FAILURE = 'CONFIG_FAILURE'
+
+export default {
+  CONFIG_REQUEST,
+  CONFIG_SUCCESS,
+  CONFIG_FAILURE
+}
