+++ src/constants/index.js
@@ -1,0 +1,10 @@
+import CONFIG from './config'
+import SESSION from './session'
+
+export const API_URL = 'http://localhost:3000/api'
+
+export default {
+  API_URL,
+  CONFIG,
+  SESSION
+}
